feat: move PluginAdvice to the model package.
with the rest of the public models.
