+++ src/functionalTest/groovy/com/autonomousapps/jvm/projects/RedundantJvmPluginsProject.groovy
@@ -3,7 +3,7 @@
 package com.autonomousapps.jvm.projects
 
 import com.autonomousapps.AbstractProject
-import com.autonomousapps.advice.PluginAdvice
+import com.autonomousapps.model.PluginAdvice
 import com.autonomousapps.kit.GradleProject
 import com.autonomousapps.kit.Source
 import com.autonomousapps.kit.SourceType
